Remove treat support (#1200)
