+++ .changeset/new-cheetahs-build.md
@@ -1,0 +1,8 @@
+---
+'braid-design-system': major
+---
+
+Remove treat support
+
+[Treat](https://seek-oss.github.io/treat/) is deprecated and has been superseded by [vanilla-extract](https://vanilla-extract.style/).
+`.treat.ts` files can no longer be used to for styling and should be converted to `.css.ts` (vanilla-extract) files.
