+++ packages/braid-design-system/css/globalHeadingStyle.docs.tsx
@@ -5,10 +5,8 @@ import { globalHeadingStyle } from '../lib/css/globalTypographyStyles';
 import source from '../lib/utils/source.macro';
 import Code from 'site/App/Code/Code';
 import type { CssDoc } from 'site/types';
-import { VanillaMigrationBanner } from './VanillaMigrationBanner';
 
 const docs: CssDoc = {
-  banner: <VanillaMigrationBanner />,
   usage: (
     <Code>{`import { globalHeadingStyle } from 'braid-design-system/css';`}</Code>
   ),
