+++ packages/braid-design-system/css/responsiveStyle.docs.tsx
@@ -5,12 +5,10 @@ import { Notice, Strong, Text, TextLink } from '../lib/components';
 import source from '../lib/utils/source.macro';
 import Code from 'site/App/Code/Code';
 import type { CssDoc } from 'site/types';
-import { VanillaMigrationBanner } from './VanillaMigrationBanner';
 
 const bps = Object.keys(breakpoints);
 
 const docs: CssDoc = {
-  banner: <VanillaMigrationBanner />,
   usage: (
     <Code>{`import { responsiveStyle } from 'braid-design-system/css';`}</Code>
   ),
