+++ packages/braid-design-system/lib/components/useToast/useToast.docs.tsx
@@ -47,7 +47,6 @@ const docs: ComponentDocs = {
             id={id}
             dedupeKey={id}
             shouldRemove={false}
-            treatTheme={theme.treatTheme}
             vanillaTheme={theme.vanillaTheme}
             onClose={() => {}}
             message="Positive toast"
@@ -210,7 +209,6 @@ const docs: ComponentDocs = {
               id={`${id}_1`}
               dedupeKey={`${id}_1`}
               shouldRemove={false}
-              treatTheme={theme.treatTheme}
               vanillaTheme={theme.vanillaTheme}
               onClose={() => {}}
               message="Positive message"
@@ -220,7 +218,6 @@ const docs: ComponentDocs = {
               id={`${id}_2`}
               dedupeKey={`${id}_2`}
               shouldRemove={false}
-              treatTheme={theme.treatTheme}
               vanillaTheme={theme.vanillaTheme}
               onClose={() => {}}
               message="Critical message"
@@ -230,7 +227,6 @@ const docs: ComponentDocs = {
               id={`${id}_3`}
               dedupeKey={`${id}_3`}
               shouldRemove={false}
-              treatTheme={theme.treatTheme}
               vanillaTheme={theme.vanillaTheme}
               onClose={() => {}}
               message="Neutral message"
@@ -240,7 +236,6 @@ const docs: ComponentDocs = {
               id={`${id}_4`}
               dedupeKey={`${id}_4`}
               shouldRemove={false}
-              treatTheme={theme.treatTheme}
               vanillaTheme={theme.vanillaTheme}
               onClose={() => {}}
               message="Neutral message with icon"
@@ -290,7 +285,6 @@ const docs: ComponentDocs = {
               id={id}
               dedupeKey={id}
               shouldRemove={false}
-              treatTheme={theme.treatTheme}
               vanillaTheme={theme.vanillaTheme}
               onClose={() => {}}
               message="Positive message"
@@ -353,7 +347,6 @@ const docs: ComponentDocs = {
               id={id}
               dedupeKey={id}
               shouldRemove={false}
-              treatTheme={theme.treatTheme}
               vanillaTheme={theme.vanillaTheme}
               onClose={() => {}}
               message="Positive message"
@@ -416,7 +409,6 @@ const docs: ComponentDocs = {
               id={id}
               dedupeKey={id}
               shouldRemove={false}
-              treatTheme={theme.treatTheme}
               vanillaTheme={theme.vanillaTheme}
               onClose={() => {}}
               tone="positive"
