+++ packages/braid-design-system/lib/playroom/FrameComponent.tsx
@@ -7,7 +7,7 @@ import {
   ToastProvider,
   useResponsiveValue,
 } from '../components';
-import type { BraidTheme } from '../themes/BraidTheme';
+import type { BraidTheme } from '../themes/makeBraidTheme';
 import { PlayroomStateProvider } from './playroomState';
 import { darkMode } from '../css/atoms/sprinkles.css';
