+++ packages/braid-design-system/lib/themes/apac/index.ts
@@ -1,7 +1,1 @@
-import { theme as treatTheme } from './theme.treat';
-import vanillaTheme from './apacTheme.css';
-
-export default {
-  ...treatTheme,
-  vanillaTheme,
-};
+export { default } from './apacTheme.css';
