+++ packages/braid-design-system/lib/themes/catho/index.ts
@@ -1,7 +1,1 @@
-import { theme as treatTheme } from './theme.treat';
-import vanillaTheme from './cathoTheme.css';
-
-export default {
-  ...treatTheme,
-  vanillaTheme,
-};
+export { default } from './cathoTheme.css';
