+++ packages/braid-design-system/lib/themes/docs/index.ts
@@ -1,7 +1,1 @@
-import { theme as treatTheme } from './theme.treat';
-import vanillaTheme from './docsTheme.css';
-
-export default {
-  ...treatTheme,
-  vanillaTheme,
-};
+export { default } from './docsTheme.css';
