+++ packages/braid-design-system/lib/themes/makeVanillaTheme.ts
@@ -63,7 +63,7 @@ const fontSizeToCapHeight = (
   };
 };
 
-export default (braidTokens: BraidTokens) => {
+export const makeVanillaTheme = (braidTokens: BraidTokens) => {
   const { name, displayName, ...tokens } = braidTokens;
   const { webFont, ...typography } = tokens.typography;
   const { foreground, background } = tokens.color;
