+++ packages/braid-design-system/lib/themes/occ/index.ts
@@ -1,7 +1,1 @@
-import { theme as treatTheme } from './theme.treat';
-import vanillaTheme from './occTheme.css';
-
-export default {
-  ...treatTheme,
-  vanillaTheme,
-};
+export { default } from './occTheme.css';
