+++ packages/braid-design-system/lib/themes/seekBusiness/index.ts
@@ -1,7 +1,1 @@
-import { theme as treatTheme } from './theme.treat';
-import vanillaTheme from './seekBusinessTheme.css';
-
-export default {
-  ...treatTheme,
-  vanillaTheme,
-};
+export { default } from './seekBusinessTheme.css';
