--- packages/braid-design-system/lib/themes/seekBusiness/theme.treat.ts
@@ -1,4 +1,0 @@
-import { makeBraidTheme } from '../makeBraidTheme';
-import tokens from './tokens';
-
-export const theme = makeBraidTheme(tokens);
