--- packages/braid-design-system/lib/themes/treatTheme.d.ts
@@ -1,5 +1,0 @@
-declare module 'treat/theme' {
-  import type { TreatTheme } from './makeBraidTheme';
-
-  export interface Theme extends TreatTheme {}
-}
